+++ model/Airbnbs.java
@@ -1,19 +1,17 @@
 package aireats.model;
 
-import java.math.BigDecimal;
-
 public class Airbnbs {
     protected String airbnbId;
     protected String name;
     protected String city;
     protected String neighborhood;
     protected String state;
-    protected BigDecimal latitude;
-    protected BigDecimal longitude;
+    protected Double latitude;
+    protected Double longitude;
 
     // Everything included
     public Airbnbs(String airbnbId, String name, String city, String neighborhood, String state,
-                  BigDecimal latitude, BigDecimal longitude) {
+                  Double latitude, Double longitude) {
         this.airbnbId = airbnbId;
         this.name = name;
         this.city = city;
@@ -31,7 +29,7 @@ public class Airbnbs {
     
     // Everything except PK
     public Airbnbs(String name, String city, String neighborhood, String state,
-            BigDecimal latitude, BigDecimal longitude) {
+    		Double latitude, Double longitude) {
     	this.name = name;
         this.city = city;
         this.neighborhood = neighborhood;
@@ -80,19 +78,19 @@ public class Airbnbs {
         this.state = state;
     }
 
-    public BigDecimal getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public BigDecimal getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
